Added a helper file for FilterFormCard component.
